Added task scroll handler and commands

+++ src/main/java/rbasamoyai/industrialwarfare/common/capabilities/itemstacks/taskscroll/ITaskScrollDataHandler.java
@@ -7,7 +7,7 @@ import rbasamoyai.industrialwarfare.common.items.taskscroll.TaskScrollOrder;
 
 public interface ITaskScrollDataHandler {
 
-	public TaskScrollOrder getOrder(byte index);
+	public TaskScrollOrder getOrder(int index);
 	
 	public void setLabel(ItemStack label);
 	public ItemStack getLabel();
